% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mission_file.R
\name{file_values_extracted}
\alias{file_values_extracted}
\title{Extract used values into list for input into .mission file FOR .PATH FILE
FOR .PLAN FILE}
\usage{
file_values_extracted(plan_file, new_coordinates)
}
\description{
Extract used values into list for input into .mission file FOR .PATH FILE
FOR .PLAN FILE
}
