% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mission_file.R
\name{gen_filled_template}
\alias{gen_filled_template}
\title{Fills a template with correct values FOR .PLAN FILE}
\usage{
gen_filled_template(list_of_filled_vals, empty_template)
}
\description{
Fills a template with correct values FOR .PLAN FILE
}
