% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_and_move.R
\name{move_to_coordinates}
\alias{move_to_coordinates}
\title{Moves geo-formatted data points to given new location (lat/long)}
\usage{
move_to_coordinates(coordinates, waypoint = 1, new_lat = NULL,
  new_lon = NULL)
}
\arguments{
\item{coordinates}{Formatted geo-coordinates.}

\item{waypoint}{The waypoint for which the coordinates are given, project moves along to that.}

\item{new_lat}{(~y), new lat for the waypoint to be moved to, if not given, unchanged}

\item{new_lon}{(~x), new long for the waypoint to be moved to, if not given, unchanged}
}
\value{
new_coord_df Same format as before function, to ensure order of rotation and movement is irrelevant.
}
\description{
Moves geo-formatted data points to given new location (lat/long)
}
