% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_mission_file.R
\name{read_in_mission_file}
\alias{read_in_mission_file}
\title{Reads in file and returns it, together with  in}
\usage{
read_in_mission_file(path)
}
\arguments{
\item{path}{Path to .mission or .plan file}
}
\value{
list containing coordinates, file-lines and extension info
}
\description{
Reads in file and returns it, together with  in
}
