% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_and_move.R
\name{rotate_geo_formatted}
\alias{rotate_geo_formatted}
\title{Rotates geo-formatted data points by given degrees}
\usage{
rotate_geo_formatted(geo_formatted_coords, degrees_rot, waypoint_center = 1)
}
\arguments{
\item{geo_formatted_coords}{Formatted geo-coordinates.}

\item{degrees_rot}{Degrees rotation}

\item{waypoint_center}{The waypoint which the rotation should pivot around (1-indexed)}
}
\value{
rotated Same format as before function, to ensure order of rotation and movement is irrelevant.
}
\description{
Rotates geo-formatted data points by given degrees
}
