% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mission_file.R
\name{write_mission_file}
\alias{write_mission_file}
\title{Writes a .mission file (currently no .plan files)}
\usage{
write_mission_file(new_coordinates, filepath, prev_file_lines, file_extension)
}
\arguments{
\item{new_coordinates}{Coordinates for the new plan after alterations}

\item{filepath}{File name/relative path+file name}

\item{prev_file_lines}{original files read in}

\item{file_extension}{extension used for read in file}
}
\description{
Writes a .mission file (currently no .plan files)
}
